feat: 管理者権限保護を実装 - パフォーマンスダッシュボードへのアクセス制限
- Userモデルにroleフィールド追加（user/admin）
- 認証セッションにrole情報を含めるように拡張
- requireAdmin()ユーティリティ関数を実装
- ダッシュボードページに管理者権限チェックを適用
- 権限がないユーザーは適切にリダイレクト

セキュリティ強化のため、パフォーマンスダッシュボードは
管理者権限を持つユーザーのみアクセス可能になりました。
