+++ prisma/schema.prisma
@@ -82,6 +82,7 @@ model User {
   name          String?
   image         String?
   password      String?
+  role          String        @default("user") // "user" | "admin"
   createdAt     DateTime      @default(now())
   updatedAt     DateTime      @updatedAt
   accounts      Account[]
